step1

+++ Mcu/K19XXVK035/Src/eeprom.c
@@ -3,7 +3,7 @@
 
 #define page_size 0x400                 
 
-void save_flash_nolib(uint8_t *data, uint16_t length, uint32_t add)
+void save_flash_nolib(uint8_t *data, int length, uint32_t add)
 {   
     volatile uint32_t start_addr;
     uint32_t data_to_FLASH[2];
